Refactor homepage component to align with updated button component and global styling; create new file for styled components
This commit updates the homepage component to match the updated button component and global styling. A new file has been created to organize styled components for improved code structure and maintainability.
